requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
botocore==1.27.41
celery==5.2.7
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
colorama==0.4.4
Django==4.0.4
django-cloudinary-storage==0.3.0
django-debug-toolbar==3.4.0
django-filter==22.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
jmespath==1.0.1
kombu==5.2.4
mccabe==0.6.1
Pillow==9.1.1
portalocker==2.5.1
prompt-toolkit==3.0.29
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
rabbitmq==0.2.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
